Leader portraits and a few small changes for Britain
Leader portraits added for Cornwall, Mann, the Clans of Scotland (got a tag change as well), The Free Government of Britain, Bowie and Fripps nations and the Free Government of Britain.

Also changed the name of Bowies nation to the "Realm of the Thin White Duke", Ulster is not quite as bright a colour as it was and the Falklands are British again.